+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/controls/DynamicScrollableList.java
@@ -34,7 +34,7 @@ import com.gitlab.cdagaming.unilib.utils.ResourceUtils;
 import com.gitlab.cdagaming.unilib.utils.gui.RenderUtils;
 import com.gitlab.cdagaming.unilib.utils.gui.controls.ScrollableListControl;
 import com.gitlab.cdagaming.unilib.utils.gui.integrations.ExtendedScreen;
-import external.io.github.classgraph.ClassInfo;
+import unilib.external.io.github.classgraph.ClassInfo;
 import io.github.cdagaming.unicore.impl.Pair;
 import io.github.cdagaming.unicore.utils.MappingUtils;
 import io.github.cdagaming.unicore.utils.StringUtils;